refactored events views to fit with new urlpy layout
