+++ ispa/events/templates/dashboard.html
@@ -1,11 +1,9 @@
 {% extends "base.html" %}
-{% load crispy_forms_tags %}
 
 {% block content %}
 <h2>events</h2>
-{% crispy event_form %}
-{% for obj in objects_list %}
-   {{obj.active_events}}
+{% for obj in active_events %}
+   {{obj}}
 {% endfor %}
 
 {% endblock %}
