--- ispa/events/views/__init__.py
@@ -1,7 +1,0 @@
-from .events import (
-    EventDetailView,
-    CreateEventView,
-    EditEventView,
-    EventListView
-)
-from .dashboard import EventDashboard
