+++ ispa/events/views/events.py
@@ -7,6 +7,21 @@ from django.views.generic import (
 from events import models
 
 
+class EventDashboard(TemplateView):
+    template_name = 'dashboard.html'
+
+    def dispatch(self, *args, **kwargs):
+        return super(EventDashboard, self).dispatch(*args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        context = super(EventDashboard, self).get_context_data(**kwargs)
+        context['active_events'] = models.Event.objects.filter(
+            is_active=True,
+        ).order_by('-name')[:5]
+
+        return context
+
+
 class EventDetailView(DetailView):
 
     model = models.Event
@@ -19,24 +34,24 @@ class EventDetailView(DetailView):
     def dispatch(self, *args, **kwargs):
         self.event = get_object_or_404(
             models.Event,
-            name=self.kwargs['name']
+            name=self.kwargs['event_slug']
         )
         return super(EventDetail, self).dispatch(*args, **kwargs)
 
     def get_context_data(self, **kwargs):
         context = super(EventDetail, self).get_context_data(**kwargs)
         try:
-            creator = models.Creator.objects.get(
+            owners = models.Owner.objects.get(
                 event=self.event,
-                owner=self.user,
+                user=self.user,
                 is_active=True
             )
-        except models.Creator.DoesNotExist:
-            creator = None
-        if creator:
-            context['creator'] = creator
+        except models.Owner.DoesNotExist:
+            owner = None
+        if owners:
+            context['owners'] = owners
 
-        context['guests'] = self.guests
+        context['event'] = self.event
         return context
 
 
@@ -96,3 +111,10 @@ class EventListView(ListView):
     @login_required
     def dispatch(self, request, *args, **kwargs):
         return super(EventList, self).dispatch(request, *args, **kwargs)
+
+
+dashboard_view = EventDashboard.as_view()
+detail_view = EventDetailView.as_view()
+create_view = CreateEventView.as_view()
+edit_view = EditEventView.as_view()
+list_view = EventListView.as_view()
